backend/requirements.txt
amqp==5.3.1
asgiref==3.11.0
billiard==4.2.4
celery==5.6.2
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cron_descriptor==2.0.6
cryptography==46.0.4
dj-database-url==3.1.0
Django==5.1.15
django-celery-beat==2.8.1
django-cors-headers==4.9.0
django-timezone-field==7.2.1
django_celery_results==2.6.0
djangorestframework==3.16.1
greenlet==3.3.1
gunicorn==25.0.0
idna==3.11
kombu==5.6.2
mozilla-django-oidc==5.0.2
packaging==26.0
playwright==1.58.0
prompt_toolkit==3.0.52
psycopg2-binary==2.9.11
pycparser==3.0
pyee==13.0.0
PyJWT==2.11.0
python-crontab==3.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
redis==7.1.0
requests==2.32.5
six==1.17.0
SQLAlchemy==2.0.46
sqlparse==0.5.5
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
urllib3==2.6.3
vine==5.1.0
wcwidth==0.5.3
